+++ packages/datagraphic/stories/DataGraphic.stories.js
@@ -1,7 +1,7 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { withKnobs } from "@storybook/addon-knobs";
 import Borders from "./Borders.svelte";
-
+import MarginsStory from "./Margins.svelte";
 import LayersStory from "./Layers.svelte";
 
 export default {
@@ -13,4 +13,8 @@ export const Layers = () => ({ Component: LayersStory });
 
 export const borders = () => ({
   Component: Borders,
+});
+
+export const Margins = () => ({
+  Component: MarginsStory,
 });